First commit of project source, developed from our UML
These classes were based upon the UML our team developed,
with some minor changes here and there.  These changes
were mostly to fix typos, but there were a few cases where
the diagram simply did not make sense, or was unclear,
and in those spots I took the liberty of fixing things.

The file TestHarness.java contains a main method with a few
simple tests to demonstrate that the code is working.

The /lib/ directory contains the json library which will
be required by the parsing code when that is added.
Thought it would be good to include it so we know everyone
is working from the same set of files.

Please feel free to tidy this code up - I'm certain that
there are inconsistancies, such as in naming conventions,
etc.